refactor(validation): share ASCII control checks across owners
