+++ packages/normalization-core/src/string-normalization.test.ts
@@ -1,6 +1,7 @@
 // Normalization Core tests cover string normalization behavior.
 import { describe, expect, it } from "vitest";
 import {
+  containsAsciiControlCharacter,
   filterStringEntries,
   normalizeAtHashSlug,
   normalizeCsvOrLooseStringList,
@@ -20,6 +21,18 @@ import {
 } from "./string-normalization.js";
 
 describe("normalization-core/string-normalization", () => {
+  it.each([
+    { label: "empty", value: "", expected: false },
+    { label: "printable Unicode", value: "fix/a&b λ", expected: false },
+    { label: "C1", value: String.fromCharCode(0x80, 0x9f), expected: false },
+    { label: "NUL", value: `branch${String.fromCharCode(0)}name`, expected: true },
+    { label: "unit separator", value: `branch${String.fromCharCode(0x1f)}`, expected: true },
+    { label: "DEL", value: `branch${String.fromCharCode(0x7f)}name`, expected: true },
+    { label: "line feed", value: "main\n", expected: true },
+  ])("detects only ASCII controls: $label", ({ value, expected }) => {
+    expect(containsAsciiControlCharacter(value)).toBe(expected);
+  });
+
   it.each([
     { value: undefined, expected: [] },
     { value: "value", expected: [] },
