+++ src/gateway/server-methods/board.github-actions.test.ts
@@ -166,6 +166,9 @@ describe("board authenticated GitHub Actions", () => {
     { perPage: 31 },
     { perPage: 1.5 },
     { workflow: "../ci.yml" },
+    { branch: "main\n" },
+    { branch: `bad${String.fromCharCode(0)}branch` },
+    { branch: `bad${String.fromCharCode(0x7f)}branch` },
     { created: "2026-02-30" },
     { status: "unknown" },
     { excludePullRequests: "true" },
